fix: restructure to suit linting and compile with import routes
Previously had `exports` configured as part of the package to ensure
that files can be imported via `@bigbite/utils/{type}/{func}` or
similar. However this could cause numerous false positives with webpack
and eslinting rules for `import/no-unresolved`. While compiling as
expected, this rule would strictly compare the route strings of an
import and the directory structure, so report an error and failure.
Imports are expected to match the directory structure. To avoid this we
need to remove the package exports and restructure by bringing files
down a level from the `src` directory, otherwise imports would have to
include that full structure; `@bigbite/utils/src/{type}/{func}`